move functions to utils and add unit tests
