+++ web/client/utils/__tests__/CoordinatesUtils-test.js
@@ -35,8 +35,11 @@ import {
     makeBboxFromOWS,
     extractCrsFromURN,
     makeNumericEPSG,
-    getPolygonFromCircle
+    getPolygonFromCircle,
+    getProjections,
+    getExtentForProjection,
 } from '../CoordinatesUtils';
+import { setConfigProp, removeConfigProp } from '../ConfigUtils';
 import Proj4js from 'proj4';
 
 describe('CoordinatesUtils', () => {
@@ -821,5 +824,27 @@ describe('CoordinatesUtils', () => {
         expect(isNearlyEqual(polygon.geometry.coordinates[0][50][1], 15.053959221823476)).toBe(true);
         expect(isNearlyEqual(polygon.geometry.coordinates[0][20][0], 39.96717142640955)).toBe(true);
         expect(isNearlyEqual(polygon.geometry.coordinates[0][20][1], 14.956343723081114)).toBe(true);
+    });
+
+    it('getProjections returns an array projections', () => {
+        let projections = getProjections();
+        expect(Array.isArray(projections)).toBe(true);
+        // 2 items because there are no projectionDefs in config
+        expect(projections.length).toBe(2);
+
+        setConfigProp('projectionDefs',  [{code: "EPSG:900913", extent: [1, 2, 3, 5]}]);
+        projections = getProjections();
+        expect(projections.length).toBe(3);
+        removeConfigProp('projectionDefs');
+    });
+
+    it('getExtentForProjection find an Extent by projection code', () => {
+        const {extent} = getExtentForProjection("EPSG:3857");
+        expect(extent.length).toEqual(4);
+
+        // returns default incase projection doesnot exist
+        const res = getExtentForProjection("EPSG:900913");
+        expect(res.extent.length).toBe(4);
+        expect(res.extent).toEqual([-20026376.39, -20048966.10, 20026376.39, 20048966.10]);
     });
 });
